pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "hotels-api"
version = "0.0.1"
requires-python = "==3.12.0"
dependencies = [
  "fastapi",
  "python-multipart",
  "email_validator",
  "pydantic-settings",
  "psycopg[binary]",
  "asyncpg",
  "sqlalchemy",
  "greenlet",
  "alembic",
  "passlib[bcrypt]",
  "uvicorn",
  "pyjwt",
  "fastapi-cache2[redis]",
  "celery",
  "flower",
  "pillow",
  "sqladmin[full]",
]

[project.optional-dependencies]
dev = [
  "isort",
  "black",
  "docformatter",
  "pyright",
  "flake8",
  "flake8-pyproject",
  "autoflake",
  "bandit",
]

[tool.isort]
profile = "black"

[tool.black]
line-length = 88

[tool.docformatter]
in-place = true
recursive = true
black = true
blank = true

[tool.flake8]
max-line-length = 88
max-complexity = 8
ignore = ["E203", "F401", "E501", "E701"]

[tool.autoflake]
recursive = true
remove-all-unused-imports = true
ignore-init-module-imports = true

[tool.bandit]
skips = ["B101"]
